refactor: memberService 추상화 제거
