+++ travel-with-me/src/test/java/com/frog/travelwithme/intergration/member/MemberIntegrationTest.java
@@ -9,8 +9,8 @@ import com.frog.travelwithme.global.security.auth.jwt.JwtTokenProvider;
 import com.frog.travelwithme.global.security.auth.userdetails.CustomUserDetails;
 import com.frog.travelwithme.intergration.BaseIntegrationTest;
 import com.frog.travelwithme.utils.ObjectMapperUtils;
-import com.frog.travelwithme.utils.StubData;
 import com.frog.travelwithme.utils.ResultActionsUtils;
+import com.frog.travelwithme.utils.StubData;
 import com.frog.travelwithme.utils.snippet.reqeust.RequestSnippet;
 import com.frog.travelwithme.utils.snippet.response.ResponseSnippet;
 import lombok.extern.slf4j.Slf4j;
